requirements.txt
asyncio
hikari
hikari-lightbulb
pyyaml
requests
thefuzz
python-Levenshtein
git+https://github.com/Rapptz/asqlite